refactor: improve storage interfaces
